+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/ChangeFeedOperationState.java
@@ -79,10 +79,6 @@ public class ChangeFeedOperationState extends FeedOperationState {
         }
     }
 
-    public void disableSplitHandling() {
-        ModelBridgeInternal.disableSplitHandling(this.options);
-    }
-
     private static Integer getEffectiveMaxItemCount(
         CosmosPagedFluxOptions pagedFluxOptions,
         CosmosChangeFeedRequestOptions changeFeedOptions) {