+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/QueryFeedOperationState.java
@@ -4,7 +4,6 @@
 package com.azure.cosmos.implementation;
 
 import com.azure.cosmos.CosmosAsyncClient;
-import com.azure.cosmos.models.CosmosChangeFeedRequestOptions;
 import com.azure.cosmos.models.CosmosQueryRequestOptions;
 import com.azure.cosmos.models.ModelBridgeInternal;