+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/models/CosmosClientTelemetryConfig.java
@@ -285,7 +285,7 @@ public final class CosmosClientTelemetryConfig {
 
                           String validTagNames = String.join(
                               ", ",
-                              (String[]) Arrays.stream(TagName.values()).map(tag -> tag.toString()).toArray());
+                              (String[]) Arrays.stream(TagName.values()).map(TagName::toString).toArray());
 
                           throw new IllegalArgumentException(
                               String.format(
@@ -299,7 +299,7 @@ public final class CosmosClientTelemetryConfig {
                   });
 
         EnumSet<TagName> newTagNames = EnumSet.noneOf(TagName.class);
-        tagNameStream.forEach(tagName -> newTagNames.add(tagName));
+        tagNameStream.forEach(newTagNames::add);
 
         this.metricTagNamesOverride = newTagNames;